Address late review of SwitchBee
