+++ homeassistant/components/switchbee/config_flow.py
@@ -5,19 +5,18 @@ import logging
 from typing import Any
 
 from switchbee.api import CentralUnitAPI, SwitchBeeError
-from switchbee.device import DeviceType
 import voluptuous as vol
 
 from homeassistant import config_entries
 from homeassistant.const import CONF_HOST, CONF_PASSWORD, CONF_USERNAME
-from homeassistant.core import HomeAssistant, callback
+from homeassistant.core import HomeAssistant
 from homeassistant.data_entry_flow import FlowResult
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.device_registry import format_mac
 
-from .const import CONF_DEFUALT_ALLOWED, CONF_DEVICES, CONF_SWITCHES_AS_LIGHTS, DOMAIN
+from .const import DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -26,7 +25,6 @@ STEP_USER_DATA_SCHEMA = vol.Schema(
         vol.Required(CONF_HOST): cv.string,
         vol.Required(CONF_USERNAME): cv.string,
         vol.Required(CONF_PASSWORD): cv.string,
-        vol.Required(CONF_SWITCHES_AS_LIGHTS, default=False): cv.boolean,
     }
 )
 
@@ -82,47 +80,6 @@ class ConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
         return self.async_show_form(
             step_id="user", data_schema=STEP_USER_DATA_SCHEMA, errors=errors
         )
-
-    @staticmethod
-    @callback
-    def async_get_options_flow(
-        config_entry: config_entries.ConfigEntry,
-    ) -> OptionsFlowHandler:
-        """Get the options flow for this handler."""
-        return OptionsFlowHandler(config_entry)
-
-
-class OptionsFlowHandler(config_entries.OptionsFlow):
-    """Handle a option flow for AEMET."""
-
-    def __init__(self, config_entry: config_entries.ConfigEntry) -> None:
-        """Initialize options flow."""
-        self.config_entry = config_entry
-
-    async def async_step_init(self, user_input=None) -> FlowResult:
-        """Handle options flow."""
-
-        if user_input is not None:
-            return self.async_create_entry(title="", data=user_input)
-
-        all_devices = [
-            DeviceType.Switch,
-            DeviceType.TimedSwitch,
-            DeviceType.GroupSwitch,
-            DeviceType.TimedPowerSwitch,
-        ]
-
-        data_schema = {
-            vol.Required(
-                CONF_DEVICES,
-                default=self.config_entry.options.get(
-                    CONF_DEVICES,
-                    CONF_DEFUALT_ALLOWED,
-                ),
-            ): cv.multi_select([device.display for device in all_devices]),
-        }
-
-        return self.async_show_form(step_id="init", data_schema=vol.Schema(data_schema))
 
 
 class CannotConnect(HomeAssistantError):
