+++ requirements_test_all.txt
@@ -33,7 +33,7 @@ PyRMVtransport==0.3.3
 PySocks==1.7.1
 
 # homeassistant.components.switchbot
-PySwitchbot==0.19.6
+PySwitchbot==0.18.22
 
 # homeassistant.components.transport_nsw
 PyTransportNSW==0.1.1
@@ -43,7 +43,7 @@ PyTransportNSW==0.1.1
 PyTurboJPEG==1.6.7
 
 # homeassistant.components.vicare
-PyViCare==2.17.0
+PyViCare==2.16.2
 
 # homeassistant.components.xiaomi_aqara
 PyXiaomiGateway==0.13.4
@@ -121,7 +121,7 @@ aiobafi6==0.7.2
 aiobotocore==2.1.0
 
 # homeassistant.components.dhcp
-aiodiscover==1.4.13
+aiodiscover==1.4.11
 
 # homeassistant.components.dnsip
 # homeassistant.components.minecraft_server
@@ -134,7 +134,7 @@ aioeafm==0.1.2
 aioeagle==1.1.0
 
 # homeassistant.components.ecowitt
-aioecowitt==2022.09.1
+aioecowitt==2022.08.3
 
 # homeassistant.components.emonitor
 aioemonitor==1.0.5
@@ -155,7 +155,7 @@ aioguardian==2022.07.0
 aioharmony==0.2.9
 
 # homeassistant.components.homekit_controller
-aiohomekit==1.5.6
+aiohomekit==1.5.1
 
 # homeassistant.components.emulated_hue
 # homeassistant.components.http
@@ -201,7 +201,7 @@ aioopenexchangerates==0.4.0
 aiopulse==0.4.3
 
 # homeassistant.components.hunterdouglas_powerview
-aiopvapi==2.0.1
+aiopvapi==2.0.0
 
 # homeassistant.components.pvpc_hourly_pricing
 aiopvpc==3.0.0
@@ -329,10 +329,7 @@ bellows==0.33.1
 bimmer_connected==0.10.2
 
 # homeassistant.components.bluetooth
-bleak-retry-connector==1.15.1
-
-# homeassistant.components.bluetooth
-bleak==0.17.0
+bleak==0.16.0
 
 # homeassistant.components.blebox
 blebox_uniapi==2.0.2
@@ -344,10 +341,10 @@ blinkpy==0.19.0
 bluemaestro-ble==0.2.0
 
 # homeassistant.components.bluetooth
-bluetooth-adapters==0.4.1
+bluetooth-adapters==0.3.4
 
 # homeassistant.components.bluetooth
-bluetooth-auto-recovery==0.3.3
+bluetooth-auto-recovery==0.3.1
 
 # homeassistant.components.bond
 bond-async==0.1.22
@@ -368,7 +365,7 @@ brunt==1.2.0
 bsblan==0.5.0
 
 # homeassistant.components.bthome
-bthome-ble==1.2.0
+bthome-ble==1.0.0
 
 # homeassistant.components.buienradar
 buienradar==1.0.5
@@ -413,9 +410,6 @@ datadog==0.15.0
 
 # homeassistant.components.metoffice
 datapoint==0.9.8
-
-# homeassistant.components.bluetooth
-dbus-fast==1.4.0
 
 # homeassistant.components.debugpy
 debugpy==1.6.3
@@ -483,7 +477,7 @@ envoy_reader==0.20.1
 ephem==4.1.2
 
 # homeassistant.components.epson
-epson-projector==0.5.0
+epson-projector==0.4.6
 
 # homeassistant.components.faa_delays
 faadelays==0.0.7
@@ -634,7 +628,7 @@ hole==0.7.0
 holidays==0.14.2
 
 # homeassistant.components.frontend
-home-assistant-frontend==20220907.0
+home-assistant-frontend==20220905.0
 
 # homeassistant.components.home_connect
 homeconnect==0.7.2
@@ -712,7 +706,7 @@ lacrosse-view==0.0.9
 laundrify_aio==1.1.2
 
 # homeassistant.components.led_ble
-led-ble==0.9.1
+led-ble==0.7.0
 
 # homeassistant.components.foscam
 libpyfoscam==1.0
@@ -763,10 +757,10 @@ mficlient==0.3.0
 micloud==0.5
 
 # homeassistant.components.mill
-mill-local==0.2.0
+mill-local==0.1.1
 
 # homeassistant.components.mill
-millheater==0.10.0
+millheater==0.9.0
 
 # homeassistant.components.minio
 minio==5.0.10
@@ -986,7 +980,7 @@ pyMetno==0.9.0
 pyRFXtrx==0.30.0
 
 # homeassistant.components.tibber
-pyTibber==0.25.2
+pyTibber==0.24.0
 
 # homeassistant.components.nextbus
 py_nextbusnext==0.1.5
@@ -1062,6 +1056,9 @@ pyefergy==22.1.1
 
 # homeassistant.components.eight_sleep
 pyeight==0.3.0
+
+# homeassistant.components.electrasmart
+pyelectra==1.0.5
 
 # homeassistant.components.everlights
 pyeverlights==0.1.0
@@ -1172,7 +1169,7 @@ pylibrespot-java==0.1.0
 pylitejet==0.3.0
 
 # homeassistant.components.litterrobot
-pylitterbot==2022.9.1
+pylitterbot==2022.8.2
 
 # homeassistant.components.lutron_caseta
 pylutron-caseta==0.13.1
@@ -1294,7 +1291,7 @@ pyruckus==0.16
 pysabnzbd==1.1.1
 
 # homeassistant.components.sensibo
-pysensibo==1.0.20
+pysensibo==1.0.19
 
 # homeassistant.components.serial
 # homeassistant.components.zha
@@ -1336,9 +1333,6 @@ pyspcwebgw==0.4.0
 
 # homeassistant.components.squeezebox
 pysqueezebox==0.6.0
-
-# homeassistant.components.switchbee
-pyswitchbee==1.4.7
 
 # homeassistant.components.syncthru
 pysyncthru==0.7.10
@@ -1463,7 +1457,7 @@ radios==0.1.1
 radiotherm==2.1.0
 
 # homeassistant.components.rainmachine
-regenmaschine==2022.09.1
+regenmaschine==2022.08.0
 
 # homeassistant.components.renault
 renault-api==0.1.11
@@ -1551,7 +1545,7 @@ soco==0.28.0
 solaredge==0.0.2
 
 # homeassistant.components.solax
-solax==0.3.0
+solax==0.2.9
 
 # homeassistant.components.honeywell
 somecomfort==0.8.0
@@ -1603,7 +1597,7 @@ subarulink==0.5.0
 sunwatcher==0.2.1
 
 # homeassistant.components.surepetcare
-surepy==0.8.0
+surepy==0.7.2
 
 # homeassistant.components.system_bridge
 systembridgeconnector==3.4.4
@@ -1628,9 +1622,6 @@ thermobeacon-ble==0.3.1
 
 # homeassistant.components.thermopro
 thermopro-ble==0.4.3
-
-# homeassistant.components.tilt_ble
-tilt-ble==0.2.3
 
 # homeassistant.components.todoist
 todoist-python==8.0.0
@@ -1692,7 +1683,7 @@ vallox-websocket-api==2.12.0
 vehicle==0.4.0
 
 # homeassistant.components.velbus
-velbus-aio==2022.9.1
+velbus-aio==2022.6.2
 
 # homeassistant.components.venstar
 venstarcolortouch==0.18
@@ -1744,7 +1735,7 @@ wolf_smartset==0.1.11
 xbox-webapi==2.0.11
 
 # homeassistant.components.xiaomi_ble
-xiaomi-ble==0.9.3
+xiaomi-ble==0.9.0
 
 # homeassistant.components.knx
 xknx==1.0.2
@@ -1761,7 +1752,7 @@ xmltodict==0.13.0
 yalesmartalarmclient==0.3.9
 
 # homeassistant.components.yalexs_ble
-yalexs-ble==1.8.1
+yalexs-ble==1.6.4
 
 # homeassistant.components.august
 yalexs==1.2.1
@@ -1782,7 +1773,7 @@ zeroconf==0.39.1
 zha-quirks==0.0.79
 
 # homeassistant.components.zha
-zigpy-deconz==0.18.1
+zigpy-deconz==0.18.0
 
 # homeassistant.components.zha
 zigpy-xbee==0.15.0
@@ -1794,10 +1785,10 @@ zigpy-zigate==0.9.2
 zigpy-znp==0.8.2
 
 # homeassistant.components.zha
-zigpy==0.50.3
+zigpy==0.50.2
 
 # homeassistant.components.zwave_js
-zwave-js-server-python==0.41.1
+zwave-js-server-python==0.41.0
 
 # homeassistant.components.zwave_me
 zwave_me_ws==0.2.4
